+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.7.0
+
 - fix: Use `LogBox` instead of `YellowBox` if possible. #989
 - fix: Don't add `DeviceContext` default integration if `enableNative` is set to `false`. #993
 - fix: Don't log "Native Sentry SDK is disabled" if `enableNativeNagger` is set to `false`. #993